Configurable lcov_branch_coverage:
This allows the lcov_branch_coverage setting
for code coverage to be configured by the calling
script, by setting the environment variable
LCOV_BRANCH_COVERAGE.

Making lines fully covered with this option
set to 1 is not practical, because function
calls which can throw have an additional invisible
branch representing the path taken when an exception
takes place.

There are other more practical ways to get full
coverage in the presence of exceptions (e.g.
tehcniques used in Boost.Beast).

By setting lcov_branch_coverage=0 it is possible to
get full coverage for a library with effort, otherwise
it is difficult to impossible.