Clear .codecov.yml and remove from templates
When it should be used the following must be met:
  - The file needs to named "codecov.yml", no other variations
  - The file needs to be in the repo root, not only during build
Those are not met, hence the file is not used by codecov.io.

Additionally codecov is (now/always?) able to use good heuristics for
applying path prefix removal as it was intended by this .codecov.yml.
This is why it did work even though the file was not used.