get_libname: Support a list of libraries in libraries.json
Currently the script expects `libraries.json` to contain a dict with the
library metadata.
However it can contain information about multiple sub-libraries,
e.g. in Boost.Utility or Boost.Functional in which case the top-level
struct is a list containing a dict for each sub-library.
It is also possible to wrap the single dict in a list with a single item
which should be equivalent to having the dict at the top-level.

Modify the script to detect such a list and work on the first entry.