Clear CCache directory on possible corruption
CCache might segfault on e.g. MacOS which can be resolved by clearing
the cache.
This might be done by deleting the GitHub Actions cache but as there is
one for each build it might become cumbersome.
So delete the folder inside the setup script when `ccache -z`
(simplest command that triggered the segfault) fails.