GHA: Allow setting defines, variant, and linkage via matrix
Pass the matrix variables
`defines`, `variant`, `link` as
`$B2_DEFINES`, `$B2_VARIANT` and `$B2_LINK`

If they are not set in the matrix the entry in the `env` dictionary will
be empty which results in using the global default value if any or not
defining it.
I.e. an empty value has no effect and we can set it even if the matrix
key is not set.